webrtc/RTCRtpReceiver-getContributingSources.https.html

<!doctype html>
<html>
<head>
  <meta http-equiv="Content-Type" content="text/html; charset=UTF-8">
  <title>RTCRtpReceiver.prototype.getContributingSources</title>
  <script src="RTCPeerConnection-helper.js"></script>
</head>
<body>
  <div id="log"></div>
  <div>
    <video id="pc2-view" muted autoplay="autoplay"></video>
  </div>

  <!-- These files are in place when executing on W3C. -->
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
  <script type="text/javascript">
  'use strict';

  async function connectAndExpectNoCsrcs(t, kind) {
    const pc1 = new RTCPeerConnection();
    t.add_cleanup(() => pc1.close());
    const pc2 = new RTCPeerConnection();
    t.add_cleanup(() => pc2.close());

    const stream = await navigator.mediaDevices.getUserMedia({[kind]:true});
    const [track] = stream.getTracks();
    t.add_cleanup(() => track.stop());
    pc1.addTrack(track, stream);

    pc2.ontrack = function(event) {
      var videoTag = document.getElementById('pc2-view');
      videoTag.srcObject = event.streams[0];
    };

    exchangeIceCandidates(pc1, pc2);
    const trackEvent = await exchangeOfferAndListenToOntrack(t, pc1, pc2);
    await exchangeAnswer(pc1, pc2);

    assert_array_equals(trackEvent.receiver.getContributingSources(), []);
  }

  promise_test(async t => {
    await connectAndExpectNoCsrcs(t, 'audio');
  }, '[audio] getContributingSources() returns an empty list in loopback call');

  promise_test(async t => {
    await connectAndExpectNoCsrcs(t, 'video');
  }, '[video] getContributingSources() returns an empty list in loopback call');
  </script>

</body>
</html>
